backend/pyproject.toml
[tool.poetry]
name = "django-photo-api"
version = "0.1.0"
description = ""
authors = ["myar <yaroslav.markelov.09@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
